Sandbox improved: added signoff feature
The sandbox.pl page now has a 2nd part to let the user signoff the patch itself.
The signoff require that the user provide the bug number, his mail and his name.
Once the signoff has been done, a mail is sent.
If there is only 1 patch attached to the bug, the script obsolete it.
If there is more than 1 patch, the script doesn't and specify in the mail that obsolete must be made manually

IMPORTANT : the WWW:Bugzilla library doesn't handle our specific status and must be tweaked to work (more doc coming soon)