Add an AJAXy script for signing off on Koha patches
=== SYNOPSIS ===

Building on the concepts of the sandbox, this script allows for users
sign off on patches in their git install by using a fixed toolbar at the
top of every page in Koha to get bug information and trigger the sign off.

In order to use this script, make it available via Apache/Plack/whatever
at /cgi-bin/koha/signoff.pl, and install WWW::Bugzilla and git-bz, then
copy the Javascript at the end of this file into intranetuserjs or
opacuserjs (but note that if you are using it via the OPAC you may need
special rewrite rules to handle it).

=== TODO ===

Make this script more asynchronous. At the moment everything is fully
synchronous, which makes it slow.

There is currently no authentication. As this script requires that the
user provide a login and password for Bugzilla, that is not as bad as it
sounds, but some sort of authentication would be good.

There is currently no mechanism for applying patches from Bugzilla, so
anything that is going to be signed off must be manually applied using
git-bz before this script can be used. The ability to apply patches off
Bugzilla would be very useful.